Split filter into smaller components
